+++ ChangeLog.md
@@ -61,6 +61,12 @@
 
   * Add better grading-instance autoscaling calculations (Matt West).
 
+  * Add student page view tracking (Matt West).
+
+  * Add predictive grader autoscaling (Matt West).
+
+  * Add links to student questions on instructor assessment instance page (Matt West).
+
   * Fix broken file upload element (Nathan Walters).
 
   * Fix broken popover and improve assessment label styles (Nathan Walters).